+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 <!--
-SPDX-FileCopyrightText: 2013 go-bench AUTHORS
+SPDX-FileCopyrightText: 2021 Eric Neidhardt
 SPDX-License-Identifier: CC-BY-4.0
 -->
 <!-- markdownlint-disable MD022 MD032 MD024-->
@@ -8,6 +8,9 @@ SPDX-License-Identifier: CC-BY-4.0
 All notable changes to this project will be documented in this file.
 
 ## Unreleased
+
+### Changed (Breaking)
+* Switched from fasthttp to net/http -> statistics from v0.2.0 are not comparable
 ### Added
 * display version when printing usage
 * Setting for additional headers